client/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
axicli @ https://cdn.evilmadscientist.com/dl/ad/public/AxiDraw_API.zip
    # via -r requirements.in
certifi==2023.5.7
    # via requests
charset-normalizer==3.1.0
    # via requests
idna==3.4
    # via requests
ink-extensions==1.2.0
    # via
    #   axicli
    #   plotink
lxml==4.9.2
    # via
    #   axicli
    #   ink-extensions
packaging==23.1
    # via plotink
plotink==1.8.0
    # via axicli
pyserial==3.5
    # via
    #   axicli
    #   plotink
requests==2.30.0
    # via axicli
urllib3==2.0.2
    # via requests
websocket-client==1.5.1
    # via -r requirements.in
